requirements.txt
exifread==3.0.0
Pillow==10.4.0
colorama==0.4.6
tkinter
